pyproject.toml
[project]
name = "disco-toolbox"
dynamic = ["version"]
description = "A toolbox for disco simulation models."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Michiel Jansen" }]
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
]
dependencies = [
  "numpy>=1.23",
  "python-graphblas~=2025.2",
  "scipy~=1.17",
]

[build-system]
requires = [
  "scikit-build-core>=0.10",
  "setuptools-scm>=8",
  "pybind11>=2.11",
]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
editable.mode = "redirect"
wheel.packages = ["src/toolbox"]

[tool.cibuildwheel]
build = "cp311-* cp312-* cp313-*"
skip = "*-win32 *-cp313t-* *-cp314t-* *-cp314-* *-musllinux_*"
test-command = "pytest -q {project}/tests"
test-extras = ["dev"]
environment = { PIP_ONLY_BINARY = "numpy", PIP_PREFER_BINARY = "1" }

[tool.cibuildwheel.windows]
archs = ["AMD64"]

[tool.cibuildwheel.linux]
archs = ["x86_64"]

[tool.scikit-build.cmake.define]
CMAKE_DOWNLOAD_TIMEOUT = 30
PYBIND11_FINDPYTHON = true

[tool.scikit-build.metadata]
version.provider = "scikit_build_core.metadata.setuptools_scm"

[tool.setuptools_scm]
write_to = "src/toolbox/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
fallback_version = "0.0.0"

[tool.setuptools.packages.find]
where = ["src"]
include = ["toolbox*"]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "mypy>=1.10",
  "twine>=5",
  "types-pyyaml>=6.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
ignore_missing_imports = true

# Baseline for the whole project
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
show_error_codes = true

# You can also add this if you want colored output:
# color_output = true

# Per-module overrides
[[tool.mypy.overrides]]
module = "toolbox.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
